Add bmemdup_ext() to utils.[hc] to fix heap-use-after-free issues.
Copying of strings, like in dstr_ncopy(), is done with one additional
byte to add a NULL byte in case the source string is not NULL-
terminated.
This leads to various reports like the following, because the extra
byte copied is read from foreign memory that can have been freed
before.

WARNING: ThreadSanitizer: heap-use-after-free (pid=20426)
  Read of size 1 at 0x7b18000816b8 by thread T13 (mutexes: write M915, write M3096, write M177183189260164400):
    # 0 memcpy gcc-12-20230121/libsanitizer/sanitizer_common/sanitizer_common_interceptors.inc:827 (libtsan.so.2+0x61e50)
    # 1 memcpy gcc-12-20230121/libsanitizer/sanitizer_common/sanitizer_common_interceptors.inc:819 (libtsan.so.2+0x61e50)
    # 2 bmemdup libobs/util/bmem.c:160 (libobs.so.0+0x117941)
    # 3 dstr_ncopy libobs/util/dstr.c:381 (libobs.so.0+0x122847)
    # 4 <null> <null> (linux-capture.so+0xae75) <== xcomp_window_name at plugins/linux-capture/xcomposite-input.c:145
    (...)

  Previous write of size 8 at 0x7b18000816b8 by thread T13:
    # 0 free gcc-12-20230121/libsanitizer/tsan/tsan_interceptors_posix.cpp:706 (libtsan.so.2+0x496ea)
    # 1 a_free libobs/util/bmem.c:84 (libobs.so.0+0x117720)
    # 2 bfree libobs/util/bmem.c:142 (libobs.so.0+0x1178aa)
    # 3 darray_free libobs/util/darray.h:56 (libobs.so.0+0x1291ff)
    # 4 free_call_children libobs/util/profiler.c:771 (libobs.so.0+0x12ceed)
    (...)

The solution to this is to add an extra function bmemdup_ext(), which
takes two arguments for the source and destination size, of which the
larger is used to bzalloc() the target and the smaller to memcpy().

Signed-off-by: Sven Eden <sven.eden@prydeworx.com>