Do not allow localtime() to be called from multiple threads.
The function localtime() returns a pointer to a static internal tm
object on success, or null pointer otherwise. The structure may be
shared between gmtime, localtime, and ctime and may be overwritten
on each invocation.

Therefore calling CurrentDateTime() or CurrentTimeString() from
multiple thread can lead to:

WARNING: ThreadSanitizer: data race (pid=15239)
  Write of size 8 at 0x7b0400004350 by thread T16:
    # 0 free libsanitizer/tsan/tsan_interceptors_posix.cpp:706 (libtsan.so.2+0x496ea)
    # 1 tzset_internal time/tzset.c:401 (libc.so.6+0xc04d8)
    # 2 CurrentTimeString[abi:cxx11]() UI/obs-app.cpp:254 (obs+0x9cf49)
    # 3 LogStringChunk UI/obs-app.cpp:298 (obs+0x9d5b6)
    (...)

  Previous write of size 8 at 0x7b0400004350 by main thread (mutexes: write M11, write M914):
    # 0 malloc libsanitizer/tsan/tsan_interceptors_posix.cpp:647 (libtsan.so.2+0x3ef36)
    # 1 __GI___strdup string/strdup.c:42 (libc.so.6+0x9a009)
    # 2 CurrentTimeString[abi:cxx11]() UI/obs-app.cpp:254 (obs+0x9cf49)
    # 3 LogStringChunk UI/obs-app.cpp:298 (obs+0x9d5b6)
    (...)

The solution is to use a local buffer with localtime_r() instead.

Note: Do not use localtime_s(), because of its implementation in
Microsoft CRT is incompatible with the C standard since it has
reversed parameter order and returns errno_t.

Signed-off-by: Sven Eden <sven.eden@prydeworx.com>