Change obs_core_video->video_fps back to plain double.
It appears that Microsoft Visual C does not yet fully support
_Atomic(<type>), leading to atomic_load_explicit() to fail with a
runtime error. Somehow the memory gets nuked and our regular
memory_order_relaxed (0) becomes an invalid large negative number.

The solution is to use local memory fencing instead. tsan is happy
with it, and it runs fine on Windows.

Signed-off-by: Sven Eden <sven.eden@prydeworx.com>