Always send stream headers
When something is wrong with the encoder, and the muxing threads fail
ffmpeg_hls_mux_data() or ffmpeg_mux_data(), ensure that the stream
headers are sent nevertheless, because obs-ffmpeg.mux might otherwise
be waiting forever in safe_read ffmpeg-mux.c:638 waiting for the
headers.

A process stuck like that can not be stopped causing endless freezing
unless the obs-ffmpeg-mux process is somehow killed.

Signed-off-by: Sven Eden <sven.eden@prydeworx.com>