Make blog() a lot more useful in debug mode and add debug_log() macro
When built with _DEBUG defined, blog now prints the source file, line
number and function name from where it was called.

There is also a `debug_log(fmt,...)` macro then, which translates to
`blog(LOG_DEBUG, ...)` putting "[debug]" between the location and the
real fmt.
This macro is special compared to plain 'blog(LOG_DEBUG,...), because
it only exists when _DEBUG is defined. In release mode it does
nothing, so it can be put anywhere without any concern about overhead
and performance in release mode.

The change to blog() being a macro now also means, that all places
where a blog preproc macro is added, the default blog macro must be
undef before creating the new one, and that this is only created when
_DEBUG is not defined, because the prefixes these defines add are
much less usefull than what the new debugging blog macro does.

To make direct use of the new macro, this commit also plasters the
sources with debug_log() call to "trace" recording starts and stops.

Signed-off-By: sven@eden-worx.com