Make full_stop() work-flow transparent.
It has been observed that recording failures (aka freezing) often
occur after something went wrong with the encoder. Unfortunately this
was not yet caught when running in a debugger, so a synchronization
problem is likely.

Plaster full_stop(), which is called when there is some encoder,
failure, with debug_log() messages so they might give a hint what is
going wrong with the emergency stopping, causing the next start to be
incomplete.

Signed-off-by: Sven Eden <sven@eden-worx.com>