Do no longer assume capture startup always succeeds
Although there are other mechanisms to detect if recordings fail,
there have been situations in which recordings are assumed as started
but then no single packet got sent. Without even the first packet,
the timeout detection on the encoder does not kick off, so let's add
some checks to the startup of the whole recording and capture
initialization to find out where the infamous recording "freeze" is
caused.

Signed-off-by: Sven Eden <sven@eden-worx.com>