Major Overhaul: Go /W4 on VS2022 and fix (almost) all warnings
Most of these warnings are harmless, and nothing bad would happen if
they continue to be ignored. But some warnings, like those about
assignments in conditionals or shadowing, can point to real desasters
only waiting to break lose.
So it is always better to turn as many warnings on as possible, and
to give the compiler no reasons to show any of them.
(This is also the difference between "OK code quality and good code
quality. Since my first "Hello World" compiled with an Aztec C 5
compiler on an Amiga 500 in 1990, I made the mistake of ignoring so
called "harmless" warnings way too often. ;-) )

Signed-off-by: Sven Eden <sven@eden-worx.com>