Use the macro PTHREAD_MUTEX_DESTROY_SAFE where it is appropriate
Doing a full lock-unlock-cycle before destroying a mutex ensures,
that no stalled or lagging thread ends up trying to use a mutex that
has already been destroyed.
In many circumstances this extra safety net is not needed, especially
when it is ensured that all other threads that might use the mutex
are already joined.
But it does not hurt either and it is always better to be safe than
sorry. Also this method makes it less likely that future development
efforts accidently create problems where there do not need to be any.

Signed-off-by: Sven Eden <sven@eden-worx.com>