Add log to os_process_pipe_destroy() and add kill to win variant
It has been observerd that the windows variant of
`os_process_pipe_destroy()` hangs infinite on
`WaitForSingleObject()`, because the muxer process won't end.
Change INFINITE to 2 minutes for now, and kill the muxer process if
the timelimit is reached. The 2 minutes are meant to give enough time
to hook the Visual Studio debugger into the process to be able to
anmalyze the situation.
Once the cause for the unwilingness of the muxer processed is found
and fixed, the timelimit should be lowered to 3 or 5 seconds.

Signed-off-by: Sven Eden <sven@eden-worx.com>